+++ src/prelude.rs
@@ -1091,7 +1091,7 @@ where
     O: ConstantFrame<Sample = T>,
     O::Size: Size<T>,
 {
-    An(Map::new(f, Routing::Arbitrary))
+    An(Map::new(f, Routing::Arbitrary(0.0)))
 }
 
 /// Keeps a signal zero centered.
@@ -2675,4 +2675,36 @@ pub fn timer<T: Float + Atomic>(shared: &Shared<T>) -> An<Timer<T>> {
 pub fn snoop<T: Float>(capacity: usize) -> (Snoop<T>, An<SnoopBackend<T>>) {
     let (snoop, backend) = Snoop::new(capacity);
     (snoop, An(backend))
+}
+
+/// Frequency domain resynthesizer. The number of inputs is `I` and the number of outputs is `O`.
+/// The window length (in samples) must be a power of two and at least four.
+/// Processes windows of input samples transformed into the frequency domain.
+/// The processing function is issued arguments (time, window).
+/// It processes frequency domain inputs into frequency domain outputs.
+/// The latency in samples is equal to window length.
+/// If any output is a copy of an input, then the input will be reconstructed exactly
+/// once all windows are overlapping, which takes `window_length` samples.
+/// -Input(s): input signals.
+/// -Output(s): processed signals.
+///
+/// ### Example: FFT Lowpass Filter
+/// ```
+/// use fundsp::prelude::*;
+/// let cutoff = 1000.0;
+/// let resynth = resynth::<U1, U1, f32, _>(1024, |_time, fft|
+///     for i in 0 .. fft.bins() {
+///         if fft.frequency(i) <= cutoff {
+///             fft.set(0, i, fft.at(0, i));
+///         }
+///     });
+/// ```
+pub fn resynth<I, O, T, F>(window_length: usize, processing: F) -> An<Resynth<I, O, T, F>>
+where
+    I: Size<T>,
+    O: Size<T>,
+    T: Float,
+    F: FnMut(f32, &mut FftWindow) + Clone + Send + Sync,
+{
+    An(Resynth::new(window_length, processing))
 }